Extend polymorphism in a single lib.rs file
- Implement multiple mathematical implementations per transform primitive.
- Implement pipeline structural polymorphism (merging and splitting).
- Implement data representation polymorphism via a Storage abstraction.
- Consolidate all logic into a single `polimorphic/src/lib.rs` file.
- Fix issues with BigInt emission and macro variable naming identified in review.
- Ensure bit-exact recovery and identical external behavior.
